+++ src/se_theory_neutral_substrate/__init__.py
@@ -1,1 +1,0 @@
-"""se_theory_neutral_substrate package."""